pyproject.toml
[project]
name = "props2yaml"
authors = [
    { name = "Jorge F. Sánchez", email = "props2yaml@jfsanchez.net" },
]
description = "Convert .properties files into .yaml"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
keywords = [
    "yaml",
    "properties",
    "converter",
]
dependencies = [
    "pyyaml",
]
dynamic = [
    "version",
]

[project.optional-dependencies]
dev = [
    "black",
]
test = [
    "pytest",
]

[build-system]
requires = [
    "setuptools>=45",
    "setuptools_scm[toml]>=7.1"
]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
version_scheme = "no-guess-dev"
local_scheme = "no-local-version"

[project.urls]
"Homepage" = "https://github.com/jfsanchez91/props2yaml"
"Bug Tracker" = "https://github.com/jfsanchez91/props2yaml/issues"

[project.scripts]
props2yaml = "props2yaml.main:main"

[tool.black]
line-length = 120

[tool.pytest.ini_options]
pythonpath = [
  ".", "src",
]
log_cli = true
log_cli_level = "ERROR"

log_file = "pytest.log"
log_file_level = "INFO"
log_file_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_file_date_format = "%Y-%m-%d %H:%M:%S"
